# Table perf improvements