--- a/platform/features/table/src/TelemetryCollection.js
+++ b/platform/features/table/src/TelemetryCollection.js
@@ -32,6 +32,7 @@ define(
          */
         function TelemetryCollection() {
             EventEmitter.call(this, arguments);
+            this.dupeCheck = false;
             this.telemetry = [];
             this.highBuffer = [];
             this.sortField = undefined;
@@ -159,7 +160,7 @@ define(
                 var startIx = _.sortedIndex(array, item, this.sortField);
                 var endIx;
 
-                if (startIx !== array.length) {
+                if (this.dupeCheck && startIx !== array.length) {
                     endIx = _.sortedLastIndex(array, item, this.sortField);
 
                     // Create an array of potential dupes, based on having the