--- a/platform/features/table/src/TelemetryCollection.js
+++ b/platform/features/table/src/TelemetryCollection.js
@@ -188,6 +188,7 @@ define(
         TelemetryCollection.prototype.add = function (items) {
             var added = items.filter(this.addOne);
             this.emit('added', added);
+            this.dupeCheck = true;
         };
 
         /**