This includes potential improvements to table performance with large datasets. (~161k rows, many duplicate values in the sort column).

There are three different changes in this PR:

A. defer dupe check
B. shortcut insertion check
C. remove format/parse cache

test case | processData elapsed
 --- | --- 
baseline | 87s 
A | 26s 
A+B | 16s 
A+B+C | 13s 

I'd consider the 3 seconds to maintain the format/parse cache a non-negligible hit; let this be a lesson to myself as to why premature optimization is bad.  The bulk of the remaining performance hit is moment date parsing.

Note that there's also 45-60 seconds of network latency loading this table.

@akhenry would love your insights on these changes before I update tests.