@larkin 2017-07-05
Failing test was due to a bug in my code, resolved that.  And moved code to TelemetryCollection.add.