@larkin 2017-07-05
# Author Checklist

1. Changes address original issue? Y (as included here)
2. Unit tests included and/or updated with changes? Y
3. Command line build passes? Y
4. Changes have been smoke-tested? Y