@VWoeltjen 2017-12-11
### Reviewer Checklist

1. Changes appear to address issue? Y
2. Appropriate unit tests included? N (performance improvement covered by existing tests)
3. Code style and in-line documentation are appropriate? Y
4. Commit messages meet standards? Y